+++ packages/remote-feature-flag-controller/src/remote-feature-flag-controller-types.ts
@@ -40,7 +40,9 @@ export type FeatureFlagScope = {
 };
 
 export type FeatureFlagScopeValue = {
-  name: string;
+  name?: string;
+  thresholdName?: string;
+  thresholdVersion?: number;
   scope: FeatureFlagScope;
   value: Json;
 };