+++ packages/remote-feature-flag-controller/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump `@metamask/controller-utils` from `^12.0.0` to `^12.1.0` ([#8774](https://github.com/MetaMask/core/pull/8774))
 
+### Fixed
+
+- Normalize object-valued feature flag wrappers so `value` contents of threshold entries are exposed at the top level while preserving the `value` property when expected by older selectors ([#8908](https://github.com/MetaMask/core/pull/8908))
+
 ## [4.2.1]
 
 ### Changed